requirements_deploy.txt
# Core dependencies
streamlit>=1.28.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# ML and NLP
torch>=1.9.0
transformers>=4.21.0
sentence-transformers==2.2.2
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0

# Vector database
chromadb==0.4.18

# Data processing
openpyxl>=3.0.0
python-dotenv==1.0.0

# Utilities
tqdm>=4.64.0
loguru>=0.7.0

# OpenAI API
openai>=0.28.0
